requirements.txt
django==5.0.5
asgiref==3.8.1
python-dotenv==1.0.1
psycopg==3.1.18
psycopg-binary==3.1.18
django-crispy-forms==2.1.0
crispy-bootstrap5==2024.2.0
django-ckeditor==6.7.0
pillow==10.3.0
django-json-widget==2.0.1
pint==0.23
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
scipy==1.13.0
matplotlib==3.8.4
uncertainties==3.1.7
celery==5.4.0
redis==5.0.4
plotly==5.22.0
